Made major changes in all layers

+++ University.RestApi/Controllers/RolesController.cs
@@ -1,10 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using University.Application.DTOs;
 using University.Application.Interfaces;
-using University.Application.Services;
 
 namespace University.RestApi.Controllers
 {
@@ -29,15 +25,11 @@ namespace University.RestApi.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<RoleDto>> Get(Guid id)
         {
-            try
-            {
-                var role = await _roleService.GetRoleByIdAsync(id);
-                return Ok(role);
-            }
-            catch (KeyNotFoundException ex)
-            {
-                return NotFound(ex.Message);
-            }
+            var role = await _roleService.GetRoleByIdAsync(id);
+            if (role is null)
+                return NotFound("Role not found");
+
+            return Ok(role);
         }
 
         [HttpPost]
@@ -51,33 +43,25 @@ namespace University.RestApi.Controllers
         public async Task<IActionResult> Update(Guid id, RoleDto roleDto)
         {
             if (id != roleDto.Id)
-            {
                 return BadRequest();
-            }
 
-            try
-            {
-                await _roleService.UpdateRolesAsync(roleDto);
-                return NoContent();
-            }
-            catch (KeyNotFoundException ex)
-            {
-                return NotFound(ex.Message);
-            }
+            var existingRole = await _roleService.GetRoleByIdAsync(id);
+            if (existingRole is null)
+                return NotFound();
+
+            await _roleService.UpdateRolesAsync(roleDto);
+            return NoContent();
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
-            try
-            {
-                await _roleService.DeleteRolesAsync(id);
-                return NoContent();
-            }
-            catch (KeyNotFoundException ex)
-            {
-                return NotFound(ex.Message);
-            }
+            var existingRole = await _roleService.GetRoleByIdAsync(id);
+            if (existingRole is null)
+                return NotFound();
+
+            await _roleService.DeleteRolesAsync(id);
+            return NoContent();
         }
     }
 }
